+++ build.clj
@@ -17,7 +17,7 @@
 
 
 (def lib 'org.scicloj/noj)
-(def version "2-beta4")
+(def version "2-beta5.1")
 (def snapshot (str version "-SNAPSHOT"))
 (def class-dir "target/classes")
 
@@ -96,3 +96,52 @@
     (dd/deploy {:installer :remote :artifact (b/resolve-path jar-file)
                 :pom-file (b/pom-path (select-keys opts [:lib :class-dir]))}))
   opts)
+
+
+
+  
+(def clojupyter-kernel-file (format "target/%s-%s-clojupyter.jar" (name lib) version))
+
+(defn create-clojupyter-kernel "Create clojupyter kernel with noj" [opts]
+  (let [basis (b/create-basis {:aliases [:clojupyter]})]
+    (b/compile-clj {:basis basis
+                    :ns-compile ['clojupyter.kernel.core 'clojupyter.cmdline]
+                    :class-dir class-dir})
+    
+    (b/uber {:uber-file clojupyter-kernel-file
+             :class-dir "target/classes"
+             ;:conflict-handlers {:default  :warn }
+             :basis basis})))
+
+
+(defn install-clojupyter-kernel "Install  clojupyter kernel in local Jupyter" [opts]
+  
+  (let [basis    (b/create-basis {:aliases [:clojupyter]})
+        cmds     (b/java-command
+                  {:basis     basis
+                   :main      'clojure.main
+                   :main-args ["-m" "clojupyter.cmdline"
+                               "install"
+                               "--ident" (str "noj-jupyter-" version)
+                               "--jarfile" clojupyter-kernel-file]})
+        {:keys [exit]} (b/process cmds)]
+    (when-not (zero? exit) (throw (ex-info "Install clojupyter kernel failed" {})))))
+  
+  
+(defn remove-clojupyter-kernel "Install  clojupyter kernel in local Jupyter" [opts]
+  (let [basis    (b/create-basis {:aliases [:clojupyter]})
+        cmds     (b/java-command
+                  {:basis     basis
+                   :main      'clojure.main
+                   :main-args ["-m" "clojupyter.cmdline"
+                               "remove-install"
+                               (str "noj-jupyter-" version)
+                               ]})
+        {:keys [exit]} (b/process cmds)]
+    (when-not (zero? exit) (throw (ex-info "Remove clojupyter kernel failed" {})))))
+  
+
+(defn replace-clojupyter-kernel "Replaces local clojupyter kernel in local Jupyter" [opts]
+     (remove-clojupyter-kernel nil)
+      (install-clojupyter-kernel nil))
+