+++ notebooks/noj_book/tablecloth_table_processing.clj
@@ -4,7 +4,6 @@
 
 ;; last change: 2025-02-02
 
-
 ;; [Tablecloth](https://scicloj.github.io/tablecloth/)
 ;; is a table processing library
 ;; inspired by the dataframe ergonomics typicall to the [R](https://www.r-project.org/)
@@ -42,7 +41,7 @@
 
 ;; ## About this tutorial
 
-;; In this tutorial, we will demonstrate the ergonomics of so-called dataset
+;; In this tutorial, we will demonstrate the ergonomics of so-called *dataset*
 ;; data strucures provided by Tablecloth. Datasets are table-like data structures,
 ;; often called data-frames in other data science platforms.
 
@@ -173,16 +172,6 @@ some-trips
                 :overflow-y :auto
                 :background "floralwhite"}}
   some-trips])
-
-;; For use in this tutorial, let us define our own customized view:
-
-(defn compact-view [dataset]
-  (kind/hiccup
-   [:div {:style {:max-width "100%"
-                  :max-height "400px"
-                  :overflow-x :auto
-                  :overflow-y :auto}}
-    dataset]))
 
 ;; ## What is a dataset?
 
@@ -345,9 +334,7 @@ some-trips
 
 ;; We can use the `tc/info` function to summarize a dataset:
 
-(-> some-trips
-    tc/info
-    compact-view)
+(tc/info some-trips)
 
 ;; ## Reading datasets
 
@@ -361,8 +348,7 @@ some-trips
 ;; First, let us read just a few rows:
 
 (->  "data/chicago-bikes/202304_divvy_tripdata.csv.gz"
-     (tc/dataset {:num-rows 3})
-     compact-view)
+     (tc/dataset {:num-rows 3}))
 
 ;; So reading a dataset is easy, but sometimes we may wish to pass a few options
 ;; to handle it a bit better.
@@ -429,12 +415,9 @@ some-trips
                                      keyword))
                     :parser-fn {"started_at" datetime-parser
                                 "ended_at"   datetime-parser}})))
-(compact-view
- trips)
+trips
 
-(-> trips
-    tc/info
-    compact-view)
+(tc/info trips)
 
 ;; It is a whole month of bike trips!
 
@@ -467,9 +450,7 @@ some-trips
 
 ;; The first few trips:
 
-(-> trips
-    tc/head
-    compact-view)
+(tc/head trips)
 
 ;; Just a few columns:
 (-> trips